Enhance displayName of ContainerQuery to make it easier to debug
In development mode React will provide warnings such as the following:

     Error: Warning: Failed propType: Required prop `someProp` was not specified in `Connect(ContainerQuery)`. Check the render method of `Connect(ContainerQuery)`.

Before this change the warning would look like the above and you would not get the name of the component that is wrapped in the ContainerQuery.  After this change you get more information like below:

     Error: Warning: Failed propType: Required prop `someProp` was not specified in `Connect(ContainerQuery(MyComponent)))`. Check the render method of `Connect(ContainerQuery(MyComponent)))`.

This allows you to see exactly which component issued the warning.